requirements.txt
flask==2.3.3
uvicorn==0.24.0
torch==2.1.0
transformers==4.35.2
pillow==10.1.0
python-multipart==0.0.6
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.1.3
python-dotenv==1.0.0
pytest==7.4.3
httpx==0.25.2
streamlit==1.28.2